[Silver II] Title: 쇠막대기, Time: 388 ms, Memory: 29328 KB -BaekjoonHub

+++ 백준/Silver/10799. 쇠막대기/쇠막대기.kt
@@ -1,4 +1,4 @@
-import java.util.Stack
+import java.util.*
 
 fun main() {
     val br = System.`in`.bufferedReader()
@@ -7,13 +7,13 @@ fun main() {
         replace("()", "1")
     }
 
-    var anwer = 0L
+    var anwer = 0
     for (char in input) {
         val c = char.toString()
         if (c == ")") {
-            var count = 0L
+            var count = 0
             while (stk.peek() != "(") {
-                count += stk.pop().toLong()
+                count += stk.pop().toInt()
             }; stk.pop()
             anwer += (count + 1)
             if (stk.contains("(")) stk.push(count.toString())
